Move `POSCatalogSyncRemoteProtocol` mock to a separate file for reuse.
